requirements.txt
streamlit>=1.28.0
yfinance>=0.2.20
ollama>=0.2.1
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
ta>=0.10.2
scikit-learn>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0